+++ MercadoPagoSDK/MercadoPagoSDK/Add_card/AddCardFlow.swift
@@ -25,10 +25,6 @@ public class AddCardFlow: NSObject, PXFlow {
     private let model = AddCardFlowModel()
     private let navigationHandler: PXNavigationHandler
 
-    // Change in Q2 when esc info comes from backend
-    private let escEnabled: Bool = true
-    private let escManager: PXESCManager
-
     //add card flow should have 'aggregator' processing mode by default
     private lazy var mercadoPagoServices = MercadoPagoServices(publicKey: "APP_USR-5bd14fdd-3807-446f-babd-095788d5ed4d", privateKey: self.accessToken)
 
@@ -43,7 +39,7 @@ public class AddCardFlow: NSObject, PXFlow {
         self.accessToken = accessToken
         self.navigationHandler = PXNavigationHandler(navigationController: navigationController)
         MPXTracker.sharedInstance.startNewSession()
-        escManager = PXESCManager(enabled: escEnabled, sessionId: MPXTracker.sharedInstance.getSessionID(), flow: "/card_association")
+        
         super.init()
         Localizator.sharedInstance.setLanguage(string: locale)
         ThemeManager.shared.saveNavBarStyleFor(navigationController: navigationController)
@@ -187,14 +183,14 @@ public class AddCardFlow: NSObject, PXFlow {
         guard let cardToken = self.model.cardToken else {
             return
         }
-        cardToken.requireESC = escEnabled
+        cardToken.requireESC = true
         self.navigationHandler.presentLoading()
 
         self.mercadoPagoServices.createToken(cardToken: cardToken, callback: { [weak self] (token) in
             guard let self = self else { return }
             self.model.tokenizedCard = token
             if let esc = token.esc {
-                self.escManager.saveESC(firstSixDigits: token.firstSixDigits, lastFourDigits: token.lastFourDigits, esc: esc)
+                PXConfiguratorManager.escProtocol.saveESC(config: PXConfiguratorManager.escConfig, firstSixDigits: token.firstSixDigits, lastFourDigits: token.lastFourDigits, esc: esc)
             }
             self.executeNextStep()
             }, failure: { [weak self] (error) in