+++ MercadoPagoSDK/MercadoPagoSDK/Flows/InitFlow/InitFlow+Services.swift
@@ -11,7 +11,7 @@ import Foundation
 extension InitFlow {
 
     func getInitSearch() {
-        let cardIdsWithEsc = initFlowModel.getESCService()?.getSavedCardIds() ?? []
+        let cardIdsWithEsc = PXConfiguratorManager.escProtocol.getSavedCardIds(config: PXConfiguratorManager.escConfig)
 
         let discountParamsConfiguration = initFlowModel.properties.advancedConfig.discountParamsConfiguration
         let flowName: String? = MPXTracker.sharedInstance.getFlowName() ?? nil
@@ -40,6 +40,20 @@ extension InitFlow {
     }
 
     func callback(_ search: PXInitDTO) {
+        /// Hack para corregir un issue cuando hay un descuento para un medio de pago particular
+        /// El nodo coupons no trae el valor de generalCoupon y cuando usa MercadoPagoCheckoutViewModel.getPaymentOptionConfigurations
+        /// se rompe todo al no encontrar el payer_costs correspondiente al coupon
+        let generalCoupon = search.generalCoupon
+        if !generalCoupon.isEmpty,
+            !search.coupons.keys.contains(generalCoupon) {
+            search.coupons[generalCoupon] = PXDiscountConfiguration(isAvailable: true)
+        }
+        if search.selectedDiscountConfiguration == nil,
+            let selectedDiscountConfiguration = search.coupons[search.generalCoupon] {
+            search.selectedDiscountConfiguration = selectedDiscountConfiguration
+        }
+        /// Fin del hack
+
         initFlowModel.updateInitModel(paymentMethodsResponse: search)
 
         //Tracking Experiments