+++ MercadoPagoSDK/MercadoPagoSDK/Models/PXPayerCost+Business.swift
@@ -22,37 +22,6 @@ extension PXPayerCost: Cellable {
         return installmentRate > 0.0 && installments > 1
     }
 
-    func hasCFTValue() -> Bool {
-        return !String.isNullOrEmpty(getCFT())
-    }
-
-    private func getLabels() -> [String: String] {
-        let prefixes: [String] = ["CFT", "TEA"]
-        var labelsDictionary: [String: String] = [:]
-        _ = labels.filter { prefixes.contains(where: $0.hasPrefix) }.flatMap { $0.components(separatedBy: "|") }.map { (label) -> String in
-            let array = label.components(separatedBy: "_")
-            if array.count == 2 {
-                labelsDictionary[array[0]] = array[1]
-            }
-            return label
-        }
-        return labelsDictionary
-    }
-
-    func getCFT(separator: String = "") -> String? {
-        let cftString = getLabels().compactMap { (key, value) -> String? in
-            if key.hasPrefix("CFT") {
-                return "\(key)\(separator) \(value)"
-            }
-            return nil
-        }.joined()
-        return cftString
-    }
-
-    func getTEAValue() -> String? {
-        return getLabels()["TEA"]
-    }
-
     func getPayerCostForTracking(isDigitalCurrency: Bool = false) -> [String: Any] {
         var installmentDic: [String: Any] = [:]
         installmentDic["quantity"] = installments