+++ CHANGELOG.md
@@ -1,4 +1,29 @@
-# v4.35.0 
+# v4.35.4 
+🚀Private Release - 4.35.4 🚀
+MercadoPagoSDKV4 - Private Version
+- Fix discounts bug
+- Fix bundle for V4
+
+# v4.35.3 
+🚀Private Release - 4.35.3 🚀
+MercadoPagoSDKV4 - Private Version
+- Create resources bundle
+- Move ESC to PXAddons
+
+# v4.35.2
+🚀Private Release - 4.35.2 🚀
+MercadoPagoSDKV4 - Private Version
+- Fix redirect bug
+
+# v4.35.1
+🚀Private Release - 4.35.1 🚀
+MercadoPagoSDKV4 - Private Version
+- Feature CFT
+- Add backURL, redirectURL and autoreturn
+- Show discounts row for consumer credits
+- Discounts row improvements
+
+# v4.35.0
 🚀Private Release - 4.35.0 🚀
 MercadoPagoSDKV4 - Private Version
 - Use MLCardFormField as Remedy textfield